Sources/requirements.txt
asgi_lifespan==2.0.0
fastapi==0.90.1
httpx==0.23.3
passlib[bcrypt]==1.7.4
pydantic[email]==1.10.4
pyhumps==3.8.0
pytest==7.2.1
tortoise-orm[asyncpg]==0.19.3
uvicorn==0.20.0
shapely==2.0.1